Fix logic of nested loop and create grid
Added a for loop that is nested within each iteration of the 'row' loop. The new loop
iterates through each row and adds the number of boxes = to 'answer. Using 'j' instead of 'i',
as a new variable, is what provided the key to getting the second loop to work. Currently
trying to deal with the issue of box overlap in the grid